core/mce: add support for decoding and handling machine checks
This provides an initial facility to decode machine checks into
human readable strings, plus a minimum amount of metadata that
a handler has to understand in order to deal with the machine
check.

For now this is only used by skiboot to make MCE reporting nicer,
and an ERAT flush recovery attempt which is more about code
coverage than really being helpful.

  ***********************************************
  Fatal MCE at 00000000300c9c0c   .memcmp+0x3c  MSR 9000000000141002
  Cause: instruction fetch TLB multi-hit error
  Effective address: 0x00000000300c9c0c
  ...

The intention is to subsequently provide an OPAL API with this
information that will enable an OS to implement a machine
independent OPAL machine check driver.

The code and data tables are derived from Linux code that I wrote,
so relicensing is okay.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>