xive/p9: Introduce XIVE_INT_ORDER
The size of the interrupt number space is constrained by the block and
index fields of the trigger data exchanged between source units and
the XIVE IC. These are respectively 4 and 28 bits, which gives us a 32
bits interrupt number space. But the XICS emulation requires 8 bits to
encode the CPPR value. The system interrupt number space is therefore
constrained to 24 bits and on a chip, to 20 bits because the XIVE
driver configures the HW to use one block per chip.

XIVE_INT_ORDER defines the size of the interrupt number space : 1M per
chip.

To control these interrupts, the driver defines in the VC BAR of the
controller a range of 384G of ESB pages giving access to 3M interrupts.
The VSD for the memory table is smaller than the index and accesses to
some ESB pages are not backed by a memory table structure. If such an
access occurred, it would result in a FIR.

It never happened but this is something to fix with a finer configuration
of the VC BAR.

Signed-off-by: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>