secvar/test: use mbedtls cflags when building the test binaries
The edk2 test file includes some mbedtls files directly, make sure that
those also include the correct mbedtls config file.

Without this, the default config file is used, which conflicts with the
version we build as part of skiboot.

As host libc includes a SIZE_MAX macro, this also changes the SIZE_MAX
macro defined in mbedtls_config.h (needed for some mbedtls functions) to
only be defined if it isn't already.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>